Started adding checklist to createGoals screen
Currently there is an array of JSON objects which make up the different tasks in the checklist {name, complete}. The issue is that react native does not detect a state change when the JSON objects inside the array are modified. I am currently working on a way to copy the array so that a state change will be detected.